@Daenyth 2022-04-14
> Is changing "appended to "prepended" in Message.addHeader actually a non-breaking change?

I'm not as familiar with the HTTP specs as other folks here, but I'm pretty sure this is a breaking change and the order matters