@Daenyth 2022-04-14
>  The order in which header fields with the same field-name are received is therefore significant to the interpretation of the combined field value, and thus a proxy MUST NOT change the order of these field values when a message is forwarded.

https://www.rfc-editor.org/rfc/rfc2616#section-4.2