@wjoel 2022-04-14
> > The order in which header fields with the same field-name are received is therefore significant to the interpretation of the combined field value, and thus a proxy MUST NOT change the order of these field values when a message is forwarded.
> 
> https://www.rfc-editor.org/rfc/rfc2616#section-4.2

That's with the same field name, though. That could be done with a `Map[String, List[String]]`, for example, right? vert.x, Netty HTTP, Undertow, and many other frameworks use maps to store headers, so I think http4s should be able to do so as well.