@armanbilge 2022-08-10
@wjoel in light of https://github.com/http4s/http4s/pull/6578#discussion_r942508403 seems the preference is to preserve header order. Can we still do some optimizations here with that restriction?