@wjoel 2022-08-11
@armanbilge it might be possible to rework this PR to preserve order while still using `ListBuffer`, at least.

But for 1.0 I'd really, really like to move away from using a `List` to represent headers, and instead use something like `Map[String, List[String]]` (or at least something map-like), and not give any guarantees about the order. I don't think lists are particularly nice for this purpose, and searching through it over and over again to find certain headers is really inefficient. And it's really inefficient to search through it to the end to append headers as well.

At work we have a lot of headers on requests that have passed through an API gateway and auth service, and I think that's only going to be more common in the future (as API gateways and microservice architectures become more popular), so I think efficiency does matter here.