@sideeffffect 2022-08-11
Please excuse a humble question from a random person on the internet:

> move away from using a `List` to represent headers, and instead use something like `Map[String, List[String]]`

Is it really _either/or_? Could we have
 * `List[Header.Raw]` as the underlying, order preserving representation of headers (or `Vector` or something else sequence-like, whatever comes up good in benchmarks)
 * `Map[String, List[String]]` (or `Map[String, Vector[String]]` or ...) as a `lazy val`, which will get computed only at the first time somebody want to search for particular header names(s)
?