@wjoel 2022-08-12
There's nothing to excuse, I'm just a random person as well. :)

The problem with keeping `List` as the underlying representation is that http4s searches for headers like `Keep-Alive` on every request, and any non-trivial use of http4s will use middlewares like CORS which searches for `Vary` on every response, so in reality that `lazy val` would practically always be computed.

In addition, if the `List` was the underlying representation, all the "add this header" operations would still be `O(n)` (and there are many of those in our middlewares).

Some of the searches could perhaps be removed, and some of the appends be aggregated in middlewares like CORS, but that's just working around the use of `List` and it would require every middleware author to be aware of the implementation of headers and the cost or various operations. The need for such work arounds is, in my opinion, an indication that `List` isn't a good choice here. If `Map` (or something with similar operation costs) was used instead, there would be less to worry about.