@sideeffffect 2022-08-12
OK, so if I understand the issue correctly, we often need to append to headers (at the end), which is O(n) with `List`. So why not then use `Vector[Header.Raw]`? And have the `Map` as a `lazy val` as I've described above.