@sideeffffect 2022-08-12
Or, as per https://www.rfc-editor.org/rfc/rfc2616#section-4.2

> _The order in which header fields with differing field names are
   received is not significant._ However, it is "good practice" to send
   general-header fields first, followed by request-header or response-
   header fields, and ending with the entity-header fields.

have just `Map[String, Vector[String]]`...
:shrug: