@sideeffffect 2022-08-12
Yep, you and the RFC have convinced me :smile_cat: So `Map[String, Vector[String]]` in the end?