@wjoel 2022-08-12
That makes two random people on the Internet who thinks that would be a good idea (at least for 1.0, where we can still make breaking changes until it's released), but I don't know if there's any consensus among http4s maintainers on this. Doesn't seem like it, but I don't know.