@armanbilge 2022-08-15
FWIW I don't see why it's _bad_ idea and it does match the spec, so I would approve such a PR targeted at `main`. I wouldn't feel confident enough to merge it though 😅