@sideeffffect 2022-09-23
Btw, maybe it should be `Map[String, NonEmptyVector[String]]`, because if the header name is in the map, the values are guaranteed to be there. We do have `NonEmptyVector` available, right?