@wjoel 2022-09-24
> Hey @wjoel, is there any way we can progress this and merge into 0.23.x? I've done some profiling and indeed it seems like adding headers is obscenely expensive sweat_smile
>
> ...
> 
> And this one tiny `putHeaders` call results in the following stack trace and approx 16% of memory allocations in the span of 5 minutes:

Yes, it's pretty wild, isn't it? Thanks for confirming what I saw all those months ago, and for reminding me of just how bad this is. :)

I'll resume work on this PR today and/or tomorrow, and try to make it pass the tests as they are, perhaps creating another PR that removes the order preservation of the headers (which is controversial to some, but not me).