@wjoel 2022-09-24
Since this no longer includes the potentially controversial reordering of headers, and passes all (non-flaky) tests, I believe it is ready for review.

I still believe the order in which headers are added shouldn't be respected by http4s (as it's not respected by the HTTP specification), and I don't think `List` is a good choice to model headers, but that can be discussed/changed in other PRs.