@CremboC 2022-09-26
hey @wjoel unfortunately I cannot run the code in the same environment I ran earlier (where I provided the flamegraph). Best second thing I could do is running the benchmark I added in https://github.com/http4s/http4s/pull/6718, together with JFR enabled `-prof jfr`. It seems like in that particular test there isn't any noticable difference in performance or memory allocations unfortunately! It's perfectly plausible there is a difference in other tests, however.