@wjoel 2022-09-26
> hey @wjoel unfortunately I cannot run the code in the same environment I ran earlier (where I provided the flamegraph). Best second thing I could do is running the benchmark I added in #6718, together with JFR enabled `-prof jfr`. It seems like in that particular test there isn't any noticable difference in performance or memory allocations unfortunately! It's perfectly plausible there is a difference in other tests, however.

No worries, your benchmarks were still somewhat useful, I found a mistake on my part and made some minor improvements with their help. It's not surprising that you didn't see much of a difference on them as they are, but it gets slightly more interesting if you add more than one `Content-Length` (then the improved check to see if it must be replaced is useful), or even just moving that one to the end seems to make a difference.

I see quite a bit of variance on those benchmarks (in part, I believe, because it makes a difference depending on which Ryzen core ends up being used), even after I upped it to 10 warmups and 10 iterations, but:
```
main, single content-length (first):
[info] Benchmark                Mode  Cnt     Score    Error  Units
[info] HeadersBench.putHeaders  avgt   10  1032.155 ± 90.169  ns/op
main, single content-length (last):
[info] Benchmark                Mode  Cnt    Score    Error  Units
[info] HeadersBench.putHeaders  avgt   10  908.060 ± 11.742  ns/op
main, two content-length (first, and last):
[info] Benchmark                Mode  Cnt     Score     Error  Units
[info] HeadersBench.putHeaders  avgt   10  1329.752 ± 107.671  ns/op

This PR, single content-length (first):
[info] Benchmark                Mode  Cnt    Score    Error  Units
[info] HeadersBench.putHeaders  avgt   10  752.709 ± 16.718  ns/op
This PR, single content-length (last):
[info] Benchmark                Mode  Cnt    Score     Error  Units
[info] HeadersBench.putHeaders  avgt   10  872.468 ± 179.501  ns/op
This PR, two content-length (first, and last):
[info] Benchmark                Mode  Cnt     Score   Error  Units
[info] HeadersBench.putHeaders  avgt   10  1185.210 ± 4.789  ns/op
```

... and test failures tell me that `.from` isn't available on Scala 2.12, so I'll have to continue this another day. I'd like to re-run a full TFB test for plaintext, as well.