@wjoel 2022-09-27
Going to have to run some more real-world(ish) benchmarks tomorrow, currently this isn't looking too great, although the lowered gc isn't bad, and might make a bigger difference on end-to-end benchmarks.
Using `++2.13 bench/jmh:run -i 10 -wi 5 -f 1 -t 1 -prof gc org.http4s.bench.HeadersBench` in sbt shell with two content-length headers (first, and last):
```
This PR
Benchmark                                             Mode  Cnt     Score     Error   Units
HeadersBench.putHeaders                               avgt   10  1511.238 ± 224.235   ns/op
HeadersBench.putHeaders:·gc.alloc.rate                avgt   10  3942.249 ± 574.347  MB/sec
HeadersBench.putHeaders:·gc.alloc.rate.norm           avgt   10  6504.000 ±   0.001    B/op
HeadersBench.putHeaders:·gc.churn.G1_Eden_Space       avgt   10  3944.447 ± 574.498  MB/sec
HeadersBench.putHeaders:·gc.churn.G1_Eden_Space.norm  avgt   10  6507.717 ±  44.131    B/op
HeadersBench.putHeaders:·gc.churn.G1_Old_Gen          avgt   10     0.011 ±   0.003  MB/sec
HeadersBench.putHeaders:·gc.churn.G1_Old_Gen.norm     avgt   10     0.019 ±   0.004    B/op
HeadersBench.putHeaders:·gc.count                     avgt   10   475.000            counts
HeadersBench.putHeaders:·gc.time                      avgt   10   313.000                ms

main:
Benchmark                                             Mode  Cnt     Score      Error   Units
HeadersBench.putHeaders                               avgt   10  1215.887 ±  246.980   ns/op
HeadersBench.putHeaders:·gc.alloc.rate                avgt   10  4957.511 ± 1022.279  MB/sec
HeadersBench.putHeaders:·gc.alloc.rate.norm           avgt   10  6528.000 ±    0.001    B/op
HeadersBench.putHeaders:·gc.churn.G1_Eden_Space       avgt   10  4967.840 ± 1008.995  MB/sec
HeadersBench.putHeaders:·gc.churn.G1_Eden_Space.norm  avgt   10  6543.204 ±   93.237    B/op
HeadersBench.putHeaders:·gc.churn.G1_Old_Gen          avgt   10     0.011 ±    0.005  MB/sec
HeadersBench.putHeaders:·gc.churn.G1_Old_Gen.norm     avgt   10     0.015 ±    0.005    B/op
HeadersBench.putHeaders:·gc.count                     avgt   10   494.000             counts
HeadersBench.putHeaders:·gc.time                      avgt   10   356.000                 ms
```

Where "first, and last" means:
```scala
  @Benchmark
  def putHeaders: Response[IO] =
    baseMessage
      .putHeaders(`Content-Length`(1000L))
      .putHeaders(`Content-Encoding`(ContentCoding.gzip))
      .putHeaders(Server(ProductId("Benchmark")))
      .putHeaders(`Content-Length`(100L))
```